@@ -1,6 +1,5 @@
 ﻿namespace NServiceBus.Heartbeat.AcceptanceTests
 {
-    using System;
     using System.Collections.Generic;
     using System.Threading;
     using System.Threading.Tasks;
@@ -32,31 +31,24 @@
             }
 
             public override Task Shutdown(CancellationToken cancellationToken = default) => Task.CompletedTask;
-#pragma warning disable CS0618
-            public override string ToTransportAddress(QueueAddress address) => transport.ToTransportAddress(address);
-#pragma warning restore CS0618
+
+            public override string ToTransportAddress(QueueAddress address) => address.ToString();
         }
 
         public InMemoryTransport() : base(TransportTransactionMode.None, true, true, true)
         {
         }
 
-#pragma warning disable 1998
-        public override async Task<TransportInfrastructure> Initialize(HostSettings hostSettings, ReceiveSettings[] receivers,
-#pragma warning restore 1998
+        public override Task<TransportInfrastructure> Initialize(HostSettings hostSettings, ReceiveSettings[] receivers,
             string[] sendingAddresses,
             CancellationToken cancellationToken = default)
         {
             var infrastructure = new InMemTransportInfrastructure(this);
             infrastructure.Initialize();
 
-            return infrastructure;
+            return Task.FromResult<TransportInfrastructure>(infrastructure);
         }
 
-        [Obsolete("Inject the ITransportAddressResolver type to access the address translation mechanism at runtime. See the NServiceBus version 8 upgrade guide for further details. Will be treated as an error from version 9.0.0. Will be removed in version 10.0.0.", false)]
-        public override string ToTransportAddress(QueueAddress address) => address.BaseAddress;
-
-        public override IReadOnlyCollection<TransportTransactionMode> GetSupportedTransactionModes() =>
-            new[] { TransportTransactionMode.None };
+        public override IReadOnlyCollection<TransportTransactionMode> GetSupportedTransactionModes() => new[] { TransportTransactionMode.None };
     }
 }
